Generator: Moved generator registering things to a separate GeneratorManager class this isolates the concerns of the Generator class, and also removes cyclic dependencies between the Generator class and its descendents.
